monitoramento_oracle/monitor_oracle_fts/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
cx_Oracle==8.3.0
idna==3.10
pathspec==0.12.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.4.0
